+++ shelfsense/src/components/products/ProductsEditModal.jsx
@@ -16,8 +16,8 @@ import Grid from '@mui/material/Grid2';
 import {Add, Delete, Remove} from '@mui/icons-material';
 import useSnackbarStore from "../../stores/useSnackbarStore.js";
 import useProductsStore from "../../stores/useProductsStore.js";
-import {deleteProduct} from '../../util/services/Products/deleteProduct.js';
-import {updateProduct} from '../../util/services/Products/updateProduct.js'
+import {deleteProduct} from '../../services/product/deleteProduct.js';
+import {updateProduct} from '../../services/product/updateProduct.js'
 import ConfirmDialog from "../confirmDialog/ConfirmDialog.jsx"
 import CloseIcon from "@mui/icons-material/Close";
