+++ shelfsense/src/pages/StatisticsPage.jsx
@@ -4,8 +4,8 @@ import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { useEffect, useState } from "react";
 import { Sidebar } from "../components/sidebar/sidebar.jsx";
-import { calculateMonthlyEarnings } from "../util/services/SalesOrders/calculateMonthlyEarnings.js";
-import { calculateMonthlyStockUsage } from "../util/services/SalesOrders/calculateMonthlyStockUsage.js";
+import { calculateMonthlyEarnings } from "../util/salesOrder/calculateMonthlyEarnings.js";
+import { calculateMonthlyStockUsage } from "../util/salesOrder/calculateMonthlyStockUsage.js";
 import { BarChart, LineChart } from '@mui/x-charts';
 import useComponentsStore from "../stores/useComponentsStore.js";
 import useProductsStore from "../stores/useProductsStore.js";
