+++ shelfsense/src/services/component/deleteComponent.js
@@ -1,4 +1,4 @@
-import {destroyStoresAndLogout} from "../Users/destroyStoresAndLogout.js";
+import {destroyStoresAndLogout} from "../../util/user/destroyStoresAndLogout.js";
 
 export const deleteComponent = async (id) => {
     const BASE_URL = `${import.meta.env.VITE_API_URL}/components/${id}`;
