+++ shelfsense/src/services/user/login.js
@@ -1,4 +1,4 @@
-import useSnackbarStore from "../../../stores/useSnackbarStore.js";
+import useSnackbarStore from "../../stores/useSnackbarStore.js";
 
 export const login = async (name, password) => {
