+++ shelfsense/src/services/user/validateSession.js
